Update LICENSE
Welcome to Complain Onchain!
We're committed to creating a positive and helpful space for feedback. By submitting feedback on our platform, you'll be minting a unique NFT that represents your contribution to improving the user experience of Base! This NFT will be automatically sent to the respective wallet associated with the subject of your feedback.

We value transparency and accountability. Your feedback, even if categorized as a "complaint," serves the purpose of constructive improvement. Through this process, we aim to foster positive interactions and meaningful resolutions within the Base community!